email to volunteers on submiting  shopping list
