+++ src/store/events/actions.js
@@ -13,32 +13,32 @@ export async function getEventsLists(dispatch, getstate) {
   }
 }
 
-export function shopping(lists) {
-  console.log(" ShoppingLists ", lists);
-  return async (dispatch, getState) => {
-    //const token = selectToken(getState());
+// export function shopping(lists) {
+//   console.log(" ShoppingLists ", lists);
+//   return async (dispatch, getState) => {
+//     //const token = selectToken(getState());
 
-    try {
-      const response = await axios.post(
-        "http://localhost:4000/shopping",
-        lists
-      );
+//     try {
+//       const response = await axios.post(
+//         "http://localhost:4000/shopping",
+//         lists
+//       );
 
-      //console.log("Auction  FORM  Response ", response.data);
-      dispatch({ type: "LISTS_SUCCESS", payload: response.data });
-      //dispatch(setMessage("success", false, null));
-      //dispatch({ type: "SUCESS_AUCTION", payload: response.data });
-    } catch (error) {
-      // console.log("AUCTUION ERROR MESSAGE message", error.response.data);
-      // console.log("AUCTUION ERROR MESSAGE message", error.message);
-      if (error.response) {
-        console.log(error.response.data.message);
-        //dispatch(setMessage("danger", true, error.response.data));
-      } else {
-        console.log("The error is ", error.message);
-        //dispatch(setMessage("danger", true, error.message));
-      }
-      //dispatch(appDoneLoading());
-    }
-  };
-}
+//       console.log("shopping  FORM  Response ", response.data);
+//       dispatch({ type: "LISTS_SUCCESS", payload: response.data });
+//       //dispatch(setMessage("success", false, null));
+//       //dispatch({ type: "SUCESS_AUCTION", payload: response.data });
+//     } catch (error) {
+//       // console.log("AUCTUION ERROR MESSAGE message", error.response.data);
+//       // console.log("AUCTUION ERROR MESSAGE message", error.message);
+//       if (error.response) {
+//         console.log(error.response.data.message);
+//         //dispatch(setMessage("danger", true, error.response.data));
+//       } else {
+//         console.log("The error is ", error.message);
+//         //dispatch(setMessage("danger", true, error.message));
+//       }
+//       //dispatch(appDoneLoading());
+//     }
+//   };
+// }
