+++ src/store/shopping/selectors.js
@@ -1,3 +1,7 @@
-export function selectShoppingDetails(reduxState) {
+export function selectShoppingLists(reduxState) {
   return reduxState.shopping.details;
+}
+
+export function selectShoppingListId(reduxState) {
+  return reduxState.shopping.shopping;
 }
